Refactor Climber Subsystem: Consolidate Constants and Remove V1_DoomSpiral Climber
- Merged V1_DoomSpiralClimberConstants into ClimberConstants for better organization.
- Removed V1_DoomSpiralClimber class, replacing its functionality with the new Climber class.
- Updated V1_DoomSpiralRobotContainer to utilize the new Climber subsystem.
- Adjusted command structure to accommodate the new Climber implementation.
