+++ src/main/java/frc/robot/commands/v1_DoomSpiral/V1_DoomSpiralCompositeCommands.java
@@ -4,15 +4,12 @@ import static edu.wpi.first.units.Units.RadiansPerSecond;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Commands;
-import edu.wpi.team190.gompeilib.core.utility.phoenix.GainSlot;
 import edu.wpi.team190.gompeilib.subsystems.drivebases.swervedrive.SwerveDrive;
 import frc.robot.commands.shared.DriveCommands;
 import frc.robot.subsystems.shared.hood.HoodConstants.HoodGoal;
 import frc.robot.subsystems.v1_DoomSpiral.V1_DoomSpiralConstants;
 import frc.robot.subsystems.v1_DoomSpiral.V1_DoomSpiralRobotState;
 import frc.robot.subsystems.v1_DoomSpiral.V1_DoomSpiralRobotState.FixedShotParameters;
-import frc.robot.subsystems.v1_DoomSpiral.climber.V1_DoomSpiralClimber;
-import frc.robot.subsystems.v1_DoomSpiral.climber.V1_DoomSpiralClimberConstants.ClimberGoal;
 import frc.robot.subsystems.v1_DoomSpiral.intake.V1_DoomSpiralIntake;
 import frc.robot.subsystems.v1_DoomSpiral.shooter.V1_DoomSpiralShooter;
 import frc.robot.subsystems.v1_DoomSpiral.spindexer.V1_DoomSpiralSpindexer;
@@ -64,21 +61,5 @@ public class V1_DoomSpiralCompositeCommands {
                 shooter.setOverrideHoodGoal(shotParameters.hoodAngle()))
             .until(shooter::atGoal),
         spindexer.setVoltage(V1_DoomSpiralSpindexerConstants.SPINDEXER_VOLTAGE));
-  }
-
-  public static Command deployClimber(V1_DoomSpiralIntake intake, V1_DoomSpiralClimber climber) {
-    return Commands.sequence(
-        intake.stow(),
-        Commands.parallel(
-            Commands.sequence(
-                intake.setRollerVoltage(-12.0),
-                intake.waitUntilIntakeAtGoal(),
-                intake.stopRoller()),
-            climber.setPositionGoal(ClimberGoal.L1_POSITION_GOAL.getPosition(), GainSlot.ZERO)));
-  }
-
-  public static Command unClimbPostAuto(V1_DoomSpiralIntake intake, V1_DoomSpiralClimber climber) {
-    return Commands.parallel(
-        intake.stow(), climber.setPositionGoal(ClimberGoal.UNCLIMB.getPosition(), GainSlot.ZERO));
   }
 }
