+++ src/main/java/frc/robot/commands/v1_DoomSpiral/autonomous/V1_DoomSpiralAutoRightTrench2Cycle.java
@@ -7,7 +7,6 @@ import edu.wpi.team190.gompeilib.subsystems.drivebases.swervedrive.SwerveDrive;
 import frc.robot.commands.shared.DriveCommands;
 import frc.robot.commands.v1_DoomSpiral.V1_DoomSpiralCompositeCommands;
 import frc.robot.subsystems.v1_DoomSpiral.V1_DoomSpiralConstants;
-import frc.robot.subsystems.v1_DoomSpiral.climber.V1_DoomSpiralClimber;
 import frc.robot.subsystems.v1_DoomSpiral.intake.V1_DoomSpiralIntake;
 import frc.robot.subsystems.v1_DoomSpiral.intake.V1_DoomSpiralIntakeConstants;
 import frc.robot.subsystems.v1_DoomSpiral.shooter.V1_DoomSpiralShooter;
@@ -18,8 +17,7 @@ public class V1_DoomSpiralAutoRightTrench2Cycle {
       SwerveDrive drive,
       V1_DoomSpiralIntake intake,
       V1_DoomSpiralShooter shooter,
-      V1_DoomSpiralSpindexer spindexer,
-      V1_DoomSpiralClimber climber) {
+      V1_DoomSpiralSpindexer spindexer) {
 
     // Create the routine and the trajectory
