+++ src/main/java/frc/robot/subsystems/v1_DoomSpiral/V1_DoomSpiralRobotContainer.java
@@ -34,14 +34,14 @@ import frc.robot.commands.shared.DriveCommands;
 import frc.robot.commands.shared.SharedCompositeCommands;
 import frc.robot.commands.v1_DoomSpiral.V1_DoomSpiralCompositeCommands;
 import frc.robot.commands.v1_DoomSpiral.autonomous.*;
+import frc.robot.subsystems.shared.climber.Climber;
+import frc.robot.subsystems.shared.climber.ClimberConstants;
 import frc.robot.subsystems.shared.fourbarlinkage.FourBarLinkageIO;
 import frc.robot.subsystems.shared.fourbarlinkage.FourBarLinkageIOSim;
 import frc.robot.subsystems.shared.fourbarlinkage.FourBarLinkageIOTalonFX;
 import frc.robot.subsystems.shared.hood.HoodIO;
 import frc.robot.subsystems.shared.hood.HoodIOTalonFX;
 import frc.robot.subsystems.shared.hood.HoodIOTalonFXSim;
-import frc.robot.subsystems.v1_DoomSpiral.climber.V1_DoomSpiralClimber;
-import frc.robot.subsystems.v1_DoomSpiral.climber.V1_DoomSpiralClimberConstants;
 import frc.robot.subsystems.v1_DoomSpiral.intake.V1_DoomSpiralIntake;
 import frc.robot.subsystems.v1_DoomSpiral.intake.V1_DoomSpiralIntakeConstants;
 import frc.robot.subsystems.v1_DoomSpiral.leds.V1_DoomSpiralCANdle;
@@ -63,7 +63,7 @@ public class V1_DoomSpiralRobotContainer implements RobotContainer {
   private SwerveDrive drive;
   private GyroIO gyroIO;
   private V1_DoomSpiralSwank swank;
-  private V1_DoomSpiralClimber climber;
+  private Climber climber;
   private V1_DoomSpiralIntake intake;
   private V1_DoomSpiralSpindexer spindexer;
   private Vision vision;
@@ -100,8 +100,8 @@ public class V1_DoomSpiralRobotContainer implements RobotContainer {
                   V1_DoomSpiralRobotState::resetPose);
           //          swank = new V1_DoomSpiralSwank(new V1_DoomSpiralSwankIOTalonFX());
           climber =
-              new V1_DoomSpiralClimber(
-                  new ArmIOTalonFX(V1_DoomSpiralClimberConstants.CLIMBER_CONSTANTS),
+              new Climber(
+                  new ArmIOTalonFX(ClimberConstants.CLIMBER_CONSTANTS),
                   gyroIO.getRoll().asSupplier());
           intake =
               new V1_DoomSpiralIntake(
@@ -169,9 +169,7 @@ public class V1_DoomSpiralRobotContainer implements RobotContainer {
                   V1_DoomSpiralRobotState::resetPose);
           //          swank = new V1_DoomSpiralSwank(new V1_DoomSpiralSwankIOTalonFXSim());
           climber =
-              new V1_DoomSpiralClimber(
-                  new ArmIOTalonFXSim(V1_DoomSpiralClimberConstants.CLIMBER_CONSTANTS),
-                  Radians::zero);
+              new Climber(new ArmIOTalonFXSim(ClimberConstants.CLIMBER_CONSTANTS), Radians::zero);
           intake =
               new V1_DoomSpiralIntake(
                   new GenericRollerIOTalonFXSim(
@@ -221,7 +219,7 @@ public class V1_DoomSpiralRobotContainer implements RobotContainer {
     }
 
     if (climber == null) {
-      climber = new V1_DoomSpiralClimber(new ArmIO() {}, Radians::zero);
+      climber = new Climber(new ArmIO() {}, Radians::zero);
     }
 
     if (intake == null) {
@@ -300,7 +298,7 @@ public class V1_DoomSpiralRobotContainer implements RobotContainer {
         .whileTrue(V1_DoomSpiralCompositeCommands.feedCommand(shooter, spindexer))
         .onFalse(V1_DoomSpiralCompositeCommands.stopShooterCommand(shooter, spindexer));
 
-    driver.x().onTrue(V1_DoomSpiralCompositeCommands.deployClimber(intake, climber).withName("x"));
+    driver.x().onTrue(SharedCompositeCommands.deployClimber(intake, climber).withName("x"));
 
     driver.y().whileTrue(climber.climbSequenceL3()).onFalse(climber.stop());
 
@@ -369,9 +367,7 @@ public class V1_DoomSpiralRobotContainer implements RobotContainer {
 
     xkeys.e8().whileTrue(climber.counterClockwiseSlow()).onFalse(climber.setVoltage(0));
 
-    xkeys
-        .e10()
-        .onTrue(V1_DoomSpiralCompositeCommands.unClimbPostAuto(intake, climber).withName("e10"));
+    xkeys.e10().onTrue(SharedCompositeCommands.unClimbPostAuto(intake, climber).withName("e10"));
 
     // Spindexer button board commands
     xkeys
@@ -453,24 +449,17 @@ public class V1_DoomSpiralRobotContainer implements RobotContainer {
 
     autoChooser.addRoutine(
         "Left Trench Simple",
-        () ->
-            V1_DoomSpiralAutoLeftTrenchSimple.getAutoRoutine(
-                drive, intake, shooter, spindexer, climber));
+        () -> V1_DoomSpiralAutoLeftTrenchSimple.getAutoRoutine(drive, intake, shooter, spindexer));
     autoChooser.addRoutine(
         "Left Trench Anti Bucks",
         () ->
-            V1_DoomSpiralAutoLeftTrenchAntiBucks.getAutoRoutine(
-                drive, intake, shooter, spindexer, climber));
+            V1_DoomSpiralAutoLeftTrenchAntiBucks.getAutoRoutine(drive, intake, shooter, spindexer));
     autoChooser.addRoutine(
         "Right Trench Simple",
-        () ->
-            V1_DoomSpiralAutoRightTrenchSimple.getAutoRoutine(
-                drive, intake, shooter, spindexer, climber));
+        () -> V1_DoomSpiralAutoRightTrenchSimple.getAutoRoutine(drive, intake, shooter, spindexer));
     autoChooser.addRoutine(
         "Depot And Back Hub",
-        () ->
-            V1_DoomSpiralAutoDepotAndBackHub.getAutoRoutine(
-                drive, intake, shooter, spindexer, climber));
+        () -> V1_DoomSpiralAutoDepotAndBackHub.getAutoRoutine(drive, intake, shooter, spindexer));
     autoChooser.addRoutine(
         "Climb",
         () -> V1_DoomSpiralAutoClimb.getAutoRoutine(drive, intake, shooter, spindexer, climber));
